servers/MAIN/requirements.txt
pip~=20.0.2
rsa~=4.8
pyasn1~=0.4.8
six~=1.15.0
gast~=0.4.0
numpy~=1.19.5
h5py~=3.1.0
requests~=2.27.1
keras~=2.6.0
PyYAML~=6.0
clang~=5.0
MarkupSafe~=2.1.1
Werkzeug~=2.1.1
Jinja2~=3.1.1
click~=8.1.2
itsdangerous~=2.1.2
tensorflow~=2.6.2
wrapt~=1.12.1
setuptools~=44.0.0
regex~=2022.3.15
wheel~=0.35
cachetools~=4.2.4
certifi~=2021.10.8
urllib3~=1.26.9
joblib~=1.1.0
Markdown~=3.3.6
oauthlib~=3.2.0
idna~=3.3
pyparsing~=3.0.8
tqdm~=4.64.0
sacremoses~=0.0.49
tensorboard~=2.6.0
flatbuffers~=1.12
astunparse~=1.6.3
termcolor~=1.1.0
tokenizers~=0.12.1
humanfriendly~=10.0
coloredlogs~=15.0.1
filelock~=3.6.0
transformers~=4.17.0
packaging~=21.3
zipp~=3.8.0
Flask~=2.1.1
pydub~=0.25.1
google-api-core==2.7.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.13.1
google-pasta==0.2.0
googleapis-common-protos==1.56.0
google-cloud-texttospeech==2.11.1
Tashaphyne==0.3.6
farasapy==0.0.14
PyArabic==0.6.14
emoji==1.7.0